removed delete all sandboxes script

+++ automation_api_scripts/delete-sandboxes-and-blueprints/delete_sandboxes_with_target_users.py
@@ -11,29 +11,7 @@ from common import add_cs_server_args, add_time_range_args, get_cloudshell_api, 
     add_target_user_arg, validate_cloudshell_users
 import constants as const
 from cloudshell.api.cloudshell_api import ReservationShortInfo
-
-parser = ArgumentParser(prog="Delete sandboxes for target users",
-                        description="Automation API script to delete sandboxes")
-add_cs_server_args(parser)
-add_time_range_args(parser)
-add_target_user_arg(parser)
-
-# unpack args
-args_dict = vars(parser.parse_args())
-server = args_dict[const.SERVER_KEY]
-user = args_dict[const.USER_KEY]
-password = args_dict[const.PASSWORD_KEY]
-from_time = args_dict[const.FROM_KEY]
-until_time = args_dict[const.UNTIL_KEY]
-target_users = args_dict[const.TARGET_USERS_KEY]
-
-api = get_cloudshell_api(server, user, password)
-validate_cloudshell_users(api, target_users)
-all_historical_sandboxes = get_all_historical_sandboxes(api, from_time, until_time)
-
-if not all_historical_sandboxes:
-    print("No historical sandboxes found. Stopping.")
-    sys.exit(0)
+from cloudshell.api.cloudshell_api import CloudShellAPISession
 
 
 def is_user_in_sandbox(sandbox: ReservationShortInfo, target_users: List[str]):
@@ -45,23 +23,51 @@ def is_user_in_sandbox(sandbox: ReservationShortInfo, target_users: List[str]):
     return False
 
 
-target_user_sandboxes = [x for x in all_historical_sandboxes if is_user_in_sandbox(x, target_users)]
-if not target_user_sandboxes:
-    print("No sandboxes found with target users associated. Stopping.")
-    sys.exit(0)
+def delete_sandboxes_with_users(api: CloudShellAPISession, target_users: List[str], from_time: str, until_time: str):
+    validate_cloudshell_users(api, target_users)
+    all_historical_sandboxes = get_all_historical_sandboxes(api, from_time, until_time)
 
-print("Deleting historical sandboxes associated with target users...")
-failed = []
-for sandbox in target_user_sandboxes:
-    print(f"Deleting sandbox '{sandbox.Id}'")
-    try:
-        api.DeleteReservation(sandbox.Id)
-    except Exception as e:
-        print(f"Error deleting sandbox '{sandbox.Id}'. Exception - \n{type(e).__name__}: {str(e)}")
-        print("===============")
-        failed.append(sandbox.Id)
+    if not all_historical_sandboxes:
+        print("No historical sandboxes found. Stopping.")
+        sys.exit(0)
 
-if failed:
-    raise Exception(f"Failed sandbox deletions:\n{json.dumps(failed, indent=4)}")
+    target_user_sandboxes = [x for x in all_historical_sandboxes if is_user_in_sandbox(x, target_users)]
+    if not target_user_sandboxes:
+        print("No sandboxes found with target users associated. Stopping.")
+        sys.exit(0)
 
-print("Delete sandboxes for target users done.")
+    print("Deleting historical sandboxes associated with target users...")
+    failed = []
+    for sandbox in target_user_sandboxes:
+        print(f"Deleting sandbox '{sandbox.Id}'")
+        try:
+            api.DeleteReservation(sandbox.Id)
+        except Exception as e:
+            print(f"Error deleting sandbox '{sandbox.Id}'. Exception - \n{type(e).__name__}: {str(e)}")
+            print("===============")
+            failed.append(sandbox.Id)
+
+    if failed:
+        raise Exception(f"Failed sandbox deletions:\n{json.dumps(failed, indent=4)}")
+
+    print("Delete sandboxes for target users done.")
+
+
+if __name__ == "__main__":
+    parser = ArgumentParser(prog="Delete sandboxes for target users",
+                            description="Automation API script to delete sandboxes")
+    add_cs_server_args(parser)
+    add_time_range_args(parser)
+    add_target_user_arg(parser)
+
+    # unpack args
+    args_dict = vars(parser.parse_args())
+    server = args_dict[const.SERVER_KEY]
+    user = args_dict[const.USER_KEY]
+    password = args_dict[const.PASSWORD_KEY]
+    from_time = args_dict[const.FROM_KEY]
+    until_time = args_dict[const.UNTIL_KEY]
+    target_users_str = args_dict[const.TARGET_USERS_KEY]
+    target_users = target_users_str.split(" ")
+
+    api = get_cloudshell_api(server, user, password)
